requirements.txt
fastapi
uvicorn
gradio
requests
pydantic
pyyaml
